requirements.txt
Django==3.1
redis==3.5.3
dj-database-url==0.5.0
psycopg2-binary==2.8.5
pytest-django==4.5.2
djangorestframework==3.12
drf-yasg==1.20.0
web3==5.28.0